pyproject.toml
[tool.poetry]
name = "comfyui-router"
version = "0.1.0"
description = "Distributed task dispatcher for comfyui clusters"
authors = ["yada <trojblue@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
celery = "^5.4.0"
msgpack = "^1.0.8"
click = "^8.1.7"
omegaconf = "^2.3.0"
gputil = "^1.4.0"
rich = "^13.8.0"
ComfyUI-EasyNodes = "*"

[tool.poetry.scripts]
crt = 'comfyui_router.cli:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"